+++ scripts/smoketest.py
@@ -6,7 +6,6 @@ import pytest
 import requests
 import pandas as pd
 import io
-import ast
 
 API_URL = "http://localhost:8000/general/v0/general"
 # NOTE(rniko): Skip inference tests if we're running on an emulated architecture
@@ -16,7 +15,7 @@ skip_inference_tests = os.getenv("SKIP_INFERENCE_TESTS", "").lower() in {"true",
 def send_document(
     filename,
     content_type,
-    strategy="fast",
+    strategy="auto",
     output_format="application/json",
     pdf_infer_table_structure="false",
 ):
@@ -87,8 +86,10 @@ def test_happy_path(example_filename, content_type):
     and some structured response
     """
     test_file = Path("sample-docs") / example_filename
+    print(f"sending {content_type}")
     json_response = send_document(test_file, content_type)
 
+    print(json_response.content)
     assert json_response.status_code == 200
     assert len(json_response.json()) > 0
     assert len("".join(elem["text"] for elem in json_response.json())) > 20
@@ -96,7 +97,7 @@ def test_happy_path(example_filename, content_type):
     csv_response = send_document(test_file, content_type, output_format="text/csv")
     assert csv_response.status_code == 200
     assert len(csv_response.text) > 0
-    df = pd.read_csv(io.StringIO(ast.literal_eval(csv_response.text)))
+    df = pd.read_csv(io.StringIO(csv_response.text))
     assert len(df) == len(json_response.json())
 
 
@@ -109,14 +110,21 @@ def test_strategy_performance():
     performance_ratio = 4
     test_file = Path("sample-docs") / "layout-parser-paper.pdf"
 
-    start_time = time.time()
+    start_time = time.monotonic()
     response = send_document(test_file, content_type="application/pdf", strategy="hi_res")
-    hi_res_time = time.time() - start_time
+    hi_res_time = time.monotonic() - start_time
     assert response.status_code == 200
 
-    start_time = time.time()
+    start_time = time.monotonic()
+    response = send_document(test_file, content_type="application/pdf", strategy="auto")
+    auto_time = time.monotonic() - start_time
+    assert response.status_code == 200
+
+    assert hi_res_time > performance_ratio * auto_time
+
+    start_time = time.monotonic()
     response = send_document(test_file, content_type="application/pdf", strategy="fast")
-    fast_time = time.time() - start_time
+    fast_time = time.monotonic() - start_time
     assert response.status_code == 200
 
     assert hi_res_time > performance_ratio * fast_time