+++ test_general/api/test_app.py
@@ -1,12 +1,10 @@
 from pathlib import Path
-from typing import List
 
 import json
 import io
 import pytest
 import re
 import requests
-import ast
 import pandas as pd
 from fastapi.testclient import TestClient
 from unstructured_api_tools.pipelines.api_conventions import get_pipeline_path
@@ -17,14 +15,6 @@ from unstructured.staging.base import convert_to_isd
 import tempfile
 
 MAIN_API_ROUTE = get_pipeline_path("general")
-
-
-def multifile_response_to_dfs(resp: requests.Response) -> List[pd.DataFrame]:
-    s = resp.text.split(',"')
-    s[0] = s[0][1:]
-    s[-1] = s[-1][:-1]
-    s[1:] = [f'"{x}' for x in s[1:]]
-    return [pd.read_csv(io.StringIO(ast.literal_eval(i))) for i in s]
 
 
 def test_general_api_health_check():
@@ -107,8 +97,8 @@ def test_general_api(example_filename, content_type):
         data={"output_format": "text/csv"},
     )
     assert csv_response.status_code == 200
-    dfs = multifile_response_to_dfs(csv_response)
-    assert len(response.json()) == len(dfs)
+    dfs = pd.read_csv(io.StringIO(csv_response.text))
+    assert len(dfs) > 0
 
 
 def test_coordinates_param():
@@ -137,8 +127,8 @@ def test_coordinates_param():
 
     # Each element should be the same except for the coordinates field
     for i in range(len(response_with_coords)):
-        assert "coordinates" in response_with_coords[i]
-        del response_with_coords[i]["coordinates"]
+        assert "coordinates" in response_with_coords[i]["metadata"]
+        del response_with_coords[i]["metadata"]["coordinates"]
         assert response_with_coords[i] == response_without_coords[i]
 
 
@@ -282,6 +272,57 @@ def test_xml_keep_tags_param():
 
             response_with_xml_tags_index += 1
             response_without_xml_tags_index += 1
+
+
+def test_include_page_breaks_param():
+    """
+    Verify that responses do not include page breaks unless requested
+    """
+    client = TestClient(app)
+    test_file = Path("sample-docs") / "layout-parser-paper-fast.pdf"
+    response = client.post(
+        MAIN_API_ROUTE,
+        files=[("files", (str(test_file), open(test_file, "rb")))],
+        data={"strategy": "fast"},
+    )
+    assert response.status_code == 200
+    response_without_page_breaks = response.json()
+
+    response = client.post(
+        MAIN_API_ROUTE,
+        files=[("files", (str(test_file), open(test_file, "rb")))],
+        data={"include_page_breaks": "true", "strategy": "fast"},
+    )
+    assert response.status_code == 200
+    response_with_page_breaks = response.json()
+
+    # The responses should have the same content except extra PageBreak objects
+    response_with_page_breaks_index, response_without_page_breaks_index = 0, 0
+    while response_with_page_breaks_index <= len(response_without_page_breaks):
+        curr_response_with_page_breaks_element = response_with_page_breaks[
+            response_with_page_breaks_index
+        ]
+        curr_response_without_page_breaks_element = response_without_page_breaks[
+            response_without_page_breaks_index
+        ]
+        if curr_response_with_page_breaks_element["type"] == "PageBreak":
+            assert curr_response_without_page_breaks_element["type"] != "PageBreak"
+
+            response_with_page_breaks_index += 1
+        else:
+            assert (
+                curr_response_without_page_breaks_element["text"]
+                == curr_response_with_page_breaks_element["text"]
+            )
+
+            response_with_page_breaks_index += 1
+            response_without_page_breaks_index += 1
+
+    last_response_with_page_breaks_element = response_with_page_breaks[
+        response_with_page_breaks_index
+    ]
+    assert last_response_with_page_breaks_element["type"] == "PageBreak"
+    assert response_without_page_breaks[-1]["type"] != "PageBreak"
 
 
 @pytest.mark.parametrize(