+++ CHANGELOG.md
@@ -1,6 +1,14 @@
-## 0.0.31-dev0
+## 0.0.32-dev0
 
+* Improve logging of params to single line json
+* Add support for `include_page_breaks` parameter
+
+## 0.0.31
+
+* Support model name as api parameter
 * Add retry parameters on fanout requests
+* Bump unstructured library to 0.8.1
+* Fix how to remove an element's coordinate information
 
 ## 0.0.30